Rename package(core_network to core.network)
