+++ core-network/src/androidTest/java/com/hanbitkang/core/network/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.hanbitkang.core_network
+package com.hanbitkang.core.network
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
