+++ core-network/src/main/java/com/hanbitkang/core/network/model/PokemonResponse.kt
@@ -1,4 +1,4 @@
-package com.hanbitkang.core_network.model
+package com.hanbitkang.core.network.model
 
 data class PokemonResponse(
     val count: Int,
